Convert Bank Peserta to Livewire with search and sorting (Fix #78)
